% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scanario.R
\name{is_production}
\alias{is_production}
\alias{is_emission_factor}
\title{Help filter different kinds of scenarios}
\usage{
is_production(x)

is_emission_factor(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
logical
}
\description{
Help filter different kinds of scenarios
}
\examples{
library(dplyr, warn.conflicts = FALSE)

filter(scenarios, is_production(variable))
filter(scenarios, is_emission_factor(variable))
}
