% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{scenario_columns}
\alias{scenario_columns}
\title{Columns in \code{scenarios}}
\usage{
scenario_columns()
}
\value{
quosures
list
}
\description{
Help manipulate the \code{scenarios} dataset.
}
\examples{
library(dplyr)

scenarios \%>\%
  mutate(new = "Will be dropped in the next line") \%>\%
  select(!!!scenario_columns())

scenarios \%>\%
  # Arrange a few rows randomly
  sample_n(5) \%>\%
  # Re-arrange rows as in `scenarios`
  arrange(!!!scenario_columns())
}
