% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{Scenarios dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56680 rows and 12 columns.
}
\usage{
scenarios
}
\description{
This dataset combines all scenarios available at
\url{https://www.transitionmonitor.com/pacta-for-banks-2020/methodology-and-supporting-materials/}.
}
\examples{
scenarios
}
\keyword{datasets}
